// src/lib/api-client/models/discovery-preview-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Domain Flow API
 * DomainFlow is a campaign-focused domain intelligence platform. The API manages personas, proxies, keyword pipelines, monitoring, and campaign execution while exposing configuration and scoring surfaces for both UI and automation clients. 
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: api@domainflow.dev
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface DiscoveryPreviewRequest
 */
export interface DiscoveryPreviewRequest {
  /**
   * Domain name generation pattern type
   * @memberof DiscoveryPreviewRequest
   */
  'patternType': 
      'prefix_variable' | 'suffix_variable' | 'both_variable'
;
  /**
   * Fixed portion of the domain name
   * @memberof DiscoveryPreviewRequest
   */
  'constantString'?: 
        
          
          string
    ;
  /**
   * Length of variable prefix
   * @memberof DiscoveryPreviewRequest
   */
  'prefixVariableLength'?: 
        
          
          number
    ;
  /**
   * Length of variable suffix
   * @memberof DiscoveryPreviewRequest
   */
  'suffixVariableLength'?: 
        
          
          number
    ;
  /**
   * Characters to use for variable portions
   * @memberof DiscoveryPreviewRequest
   */
  'characterSet': 
        
          
          string
    ;
  /**
   * Top-level domain
   * @memberof DiscoveryPreviewRequest
   */
  'tld': 
        
          
          string
    ;
}