// src/lib/api-client/models/discovery-preview-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Domain Flow API
 * DomainFlow is a campaign-focused domain intelligence platform. The API manages personas, proxies, keyword pipelines, monitoring, and campaign execution while exposing configuration and scoring surfaces for both UI and automation clients. 
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: api@domainflow.dev
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DiscoveryPreviewResponseData } from './discovery-preview-response-data';

/**
 * Preview of discovery configuration with lineage information
 * @export
 * @interface DiscoveryPreviewResponse
 */
export interface DiscoveryPreviewResponse {
  /**
   * 
   * @memberof DiscoveryPreviewResponse
   */
  'success': 
        
          
          boolean
    ;
  /**
   * 
   * @memberof DiscoveryPreviewResponse
   */
  'requestId'?: 
        
          
          string
    ;
  /**
   * 
   * @memberof DiscoveryPreviewResponse
   */
  'data': 
        
          
          DiscoveryPreviewResponseData
    ;
}